change model size field to PositiveIntegerField

+++ listings/models.py
@@ -85,7 +85,7 @@ class PaymentSchedule(models.TextChoices):
 
 
 class SolarSolution(TimeStampedModel):
-    size = models.DecimalField(max_digits=5, decimal_places=2, help_text="Solution size in kW (e.g., 5.00 kW)")
+    size = models.PositiveIntegerField(help_text="Solution size in kW (e.g., 5.00 kW)")
     price = models.DecimalField(max_digits=10, decimal_places=2, help_text="Total price of the solution/package")
     solution_type = models.CharField(max_length=50, choices=SolutionType.choices,
                                      help_text="Select the type of solar solution")
